[Build ] registerVerifier.
